Prevent integer multiplication on log size argument.

+++ main/realmain.cc
@@ -168,8 +168,10 @@ int realmain(int argc, char *argv[]) {
     }
     if (!opts.debugLogFile.empty()) {
         // LSP could run for a long time. Rotate log files, and trim at 5 GiB. Keep around 3 log files.
+        // Cast first number to size_t to prevent integer multiplication.
         // TODO(jvilk): Reduce size once LSP logging is less chunderous.
-        auto fileSink = make_shared<spdlog::sinks::rotating_file_sink_mt>(opts.debugLogFile, 5 * 1024 * 1024 * 1024, 3);
+        auto fileSink =
+            make_shared<spdlog::sinks::rotating_file_sink_mt>(opts.debugLogFile, ((size_t)5) * 1024 * 1024 * 1024, 3);
         fileSink->set_level(spd::level::debug);
         { // replace console & fatal loggers
             vector<spd::sink_ptr> sinks{stderrColorSink, fileSink};
